did parts 4.18-23

+++ osa_4/blog/controllers/bloglist.js
@@ -1,29 +1,35 @@
 const bloglistRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
+const {userExtractor}= require('../middleware/middleware')
+
+//Note to self: om man lägger till den såhär så är den aktiv för alla sitationer
+//bloglistRouter.use(authorize)
 
 bloglistRouter.get('/', async (_request, response) => {
   const blogs = await Blog.find({}).populate('user',{username: 1, name:1, id:1})
   response.json(blogs.map(b => b.toJSON()))
 })
 
-bloglistRouter.post('/', async (request, response) => {
+bloglistRouter.post('/', userExtractor, async (request, response) => {
   const blog = request.body
+  const user = request.user
   if (!blog.url || !blog.title) {
-    response.status(400).end()
+    return response.status(400).end()
   } else {
     blog.likes = blog.likes || 0
+    if (blog.userId !== request.user.id) {
+      throw Error('Unauthorized')
+    }
     try {
-      const user = await User.findById(blog.userId)
       delete blog.userId
       blog.user = user.id
       const savedBlog = await new Blog(blog).save()
       //jo, bra o bli o fundera på det här i evighteter... man måste alltså lägga till en referense till bloggarn i användarens info 
       user.blogs = user.blogs.concat(savedBlog._id)
       await user.save()
-      response.status(201).json(savedBlog)
+      return response.status(201).json(savedBlog)
     } catch (error) {
-      response.status(401).json(error.message)
+      return response.status(401).json(error.message)
     }
   }
 })
@@ -37,7 +43,7 @@ bloglistRouter.get('/:id', async (request, response) => {
   }
 })
 
-bloglistRouter.put('/:id', async (request, response, next) => {
+bloglistRouter.put('/:id', async (request, response) => {
   const id = request.params.id
   const body = request.body
   //const user = await User.findById(body.userId)
@@ -56,8 +62,15 @@ bloglistRouter.put('/:id', async (request, response, next) => {
   }    
 })
 
-bloglistRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+bloglistRouter.delete('/:id', userExtractor, async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  if (blog.user.toString() === request.user.id) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(401).end()
+  }
 })
+
 module.exports = bloglistRouter
+
